requirements.txt
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
opencv-python==3.4.4.19
psutil==5.7.3
Shapely==1.7.1
tensorboard==1.13.1
tensorflow_cpu==1.13.1
